Refactor TokenService to separate service/data storage
