+++ server/src/main/java/org/exoplatform/chat/services/TokenServiceImpl.java
@@ -17,7 +17,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.exoplatform.chat.services.mongodb;
+package org.exoplatform.chat.services;
 
 import com.mongodb.BasicDBObject;
 import com.mongodb.DB;
@@ -36,6 +36,7 @@ import org.exoplatform.container.PortalContainer;
 import org.exoplatform.services.user.UserStateService;
 
 import javax.enterprise.context.ApplicationScoped;
+import javax.inject.Inject;
 import javax.inject.Named;
 import java.util.HashMap;
 import java.util.List;
@@ -49,21 +50,15 @@ public class TokenServiceImpl implements TokenService
   //TODO this service will not be available in 2 mode servers, we will need to find another solution (REST API /state ?)
   private UserStateService userStateService;
 
+  @Inject
+  private TokenStorage storage;
+
   public TokenServiceImpl() {
     userStateService = PortalContainer.getInstance().getComponentInstanceOfType(UserStateService.class);
   }
 
   public TokenServiceImpl(UserStateService userStateService) {
     this.userStateService = userStateService;
-  }
-
-  private DB db(String dbName)
-  {
-    if (StringUtils.isEmpty(dbName)) {
-      return ConnectionManager.getInstance().getDB();
-    } else {
-      return ConnectionManager.getInstance().getDB(dbName);
-    }
   }
 
   public String getToken(String user)
@@ -76,40 +71,12 @@ public class TokenServiceImpl implements TokenService
 
   public boolean hasUserWithToken(String user, String token, String dbName)
   {
-    DBCollection coll = db(dbName).getCollection(M_USERS_COLLECTION);
-    BasicDBObject query = new BasicDBObject();
-    query.put("user", user);
-    query.put("token", token);
-    DBCursor cursor = coll.find(query);
-    return (cursor.hasNext());
+    return storage.hasUserWithToken(user, token, dbName);
   }
 
   public void addUser(String user, String token, String dbName)
   {
-    if (!hasUserWithToken(user, token, dbName))
-    {
-      DBCollection coll = db(dbName).getCollection(M_USERS_COLLECTION);
-
-      BasicDBObject query = new BasicDBObject();
-      query.put("user", user);
-      DBCursor cursor = coll.find(query);
-      if (cursor.hasNext())
-      {
-        DBObject doc = cursor.next();
-        doc.put("token", token);
-        doc.put("isDemoUser", user.startsWith(ANONIM_USER));
-        coll.save(doc, WriteConcern.SAFE);
-      }
-      else
-      {
-        BasicDBObject doc = new BasicDBObject();
-        doc.put("_id", user);
-        doc.put("user", user);
-        doc.put("token", token);
-        doc.put("isDemoUser", user.startsWith(ANONIM_USER));
-        coll.insert(doc);
-      }
-    }
+    storage.addUser(user, token, dbName);
   }
 
   public Map<String, UserBean> getActiveUsersFilterBy(String user, String dbName, boolean withUsers, boolean withPublic, boolean isAdmin)
@@ -121,43 +88,7 @@ public class TokenServiceImpl implements TokenService
   {
     List<String> onlineUsers = userStateService.online().stream().map(u -> u.getUserId()).collect(Collectors.toList());
 
-    HashMap<String, UserBean> users = new HashMap<>();
-    DBCollection coll = db(dbName).getCollection(M_USERS_COLLECTION);
-    BasicDBObject query = new BasicDBObject();
-    if (isAdmin)
-    {
-      if (withPublic && !withUsers)
-      {
-        query.put("isDemoUser", true);
-      }
-      else if (!withPublic && withUsers)
-      {
-        query.put("isDemoUser", false);
-      }
-    }
-    else
-    {
-      query.put("isDemoUser", user.startsWith(ANONIM_USER));
-    }
-    if (limit<0) limit=0;
-    DBCursor cursor = coll.find(query).limit(limit);
-    while (cursor.hasNext())
-    {
-      DBObject doc = cursor.next();
-      String target = doc.get("user").toString();
-      // Exclude current user and not online users
-      if (!user.equals(target) && onlineUsers.contains(target)) {
-        UserBean userBean = new UserBean();
-        userBean.setName(target);
-        if (doc.get("fullname")!=null)
-          userBean.setFullname( doc.get("fullname").toString() );
-        if (doc.get("status")!=null)
-          userBean.setStatus(doc.get("status").toString());
-        users.put(target, userBean);
-      }
-    }
-
-    return users;
+    return storage.getActiveUsersFilterBy(user, onlineUsers, dbName, withUsers, withPublic, isAdmin, limit);
   }
 
   public boolean isUserOnline(String user, String dbName)
